WOW!frequency
Main focus on Fundamental frequency Detection.

Works almost perfectly.
The user can press on any fret of any string and know which out of the 12 notes in the chromatic scale is played.

For Octave separation, however you need to recheck, eg. Open B is B4 but it shows B5, etc.

More or less, frequency is detected via autocorrelation algorithm.